// Plan/react/dashboard/src/dataHooks/model/graph/performance/PerformanceGraphId.ts
export enum PerformanceGraphId {
    PLAYERS_ONLINE = "players-online",
    TPS = 'tps',
    MSPT_AVG = 'mspt-avg',
    MSPT_95TH = 'mspt-95th',
    MSPT_JITTER_AVG = 'mspt-jitter-avg',
    MSPT_JITTER_MAX = 'mspt-jitter-max',
    CHUNKS = 'chunks',
    ENTITIES = 'entities',
    CPU = 'cpu',
    RAM = 'ram',
    DISK = 'disk',
    PING = 'ping',
    PING_MIN = 'ping-min',
    PING_AVG = 'ping-avg',
    PING_MAX = 'ping-max'
}